requirements.txt
blis==0.7.4
catalogue==1.0.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
cymem==2.0.5
decorator==4.4.2
en-core-web-sm==2.3.1
ffmpeg-python==0.2.0
future==0.18.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
joblib==1.0.0
kiwisolver==1.3.1
matplotlib==3.3.3
moviepy==1.0.3
murmurhash==1.0.5
nltk==3.5
numpy==1.19.5
opencv-python==4.5.1.48
Pillow==8.1.0
plac==1.1.3
preshed==3.0.5
proglog==0.1.9
pydub==0.24.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
regex==2020.11.13
requests==2.25.1
six==1.15.0
spacy==2.3.5
spaw==0.2
srsly==1.0.5
textblob==0.15.3
thinc==7.4.5
tinydb==4.3.0
tqdm==4.56.0
urllib3==1.26.2
wasabi==0.8.1
